Minor Update
* Squash minor bug fixes (#5)
* Use `sys.exit()` calls (#4)
* Formatted python scripts with autopep8.
* Formatted bash scripts
* Added DeepSource issue badge.
* Other miscellaneous changes

+++ bash/nextcloud_snapshot.sh
@@ -4,30 +4,30 @@
 # Can be used as a Cronjob
 # Tested on Ubuntu Server 20.04 LTS
 
-# Variables
-source_path="/path/to/nextcloud" # Where nextcloud is
-archive_path="/path/to/Backups" # Where you want to store the archive
-archive_password="<Archive password goes here>"
-remote_path="google_drive:snapshots" # rClone must be configured
+# Make sure enviroinment variables are set in .bashrc
+source_path=$NC_SOURCE_PATH # Where nextcloud is
+archive_path=$NC_ARCHIVE_PATH  # Where you want to store the archive
+archive_password=$NC_ARCHIVE_PASSWORD
+remote_path=$NC_REMOTE_PATH # rClone must be configured (eg: gdrive:snapshots)
 
-start_time=`date +%s`
+start_time=$(date +%s)
 
 # Variables for pretty printing
-RED=`tput bold``tput setaf 1` # Red Color
-GREEN=`tput bold``tput setaf 2` # Green Color
-NC=`tput sgr0` # No Color
+RED=$(tput bold)$(tput setaf 1)   # Red Color
+GREEN=$(tput bold)$(tput setaf 2) # Green Color
+NC=$(tput sgr0)                   # No Color
 
 # Checking root/sudo permissions
-if [ "$(id -u)" -ne "0" ]; then 
+if [ "$(id -u)" -ne "0" ]; then
 	echo "${RED}Please run as root${NC}"
 	exit
 fi
 
-backup_dir=$(date +'%d-%m-%Y');
-archive_name="Nextcloud_Backup_${backup_dir}.7z";
+backup_dir=$(date +'%d-%m-%Y')
+archive_name="Nextcloud_Backup_${backup_dir}.7z"
 
 # Deleting previous backup file(s)
-rm -rf $archive_path/*
+rm -rf "${archive_path:?}/"*
 
 # 7z compress and password protect file
 # > sudo apt install p7zip-full
@@ -40,6 +40,6 @@ echo -e "${GREEN}Upload started....${NC}"
 rclone copy --update --verbose --transfers 1 --checkers 8 --contimeout 60s --timeout 300s --retries 3 --low-level-retries 10 --stats 1s "$archive_path" "$remote_path"
 echo -e "${GREEN}Finished uploading${NC}"
 
-end_time=`date +%s`
-runtime=$((end_time-start_time))
+end_time=$(date +%s)
+runtime=$((end_time - start_time))
 echo -e "${GREEN}Completed in ${runtime} seconds ${NC}"
